[sourcegen] Eliminate csharp/_Config.py
